Add Gate

+++ app/Http/Controllers/UserController.php
@@ -14,6 +14,7 @@ use Illuminate\Database\Eloquent\Model;
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Hash;
 use Symfony\Component\HttpFoundation\Response;
 
@@ -27,6 +28,8 @@ class UserController extends Controller
      */
     public function index(): AnonymousResourceCollection
     {
+        Gate::authorize('view', 'users');
+
         return UserResource::collection(User::paginate());
     }
 
@@ -37,6 +40,8 @@ class UserController extends Controller
      */
     public function show(int $id): UserResource
     {
+        Gate::authorize('view', 'users');
+
         return new UserResource(User::find($id));
     }
 
@@ -46,6 +51,8 @@ class UserController extends Controller
      */
     public function store(UserCreateRequest $request): Response
     {
+        Gate::authorize('edit', 'users');
+
         $user = User::create(
             $request->only('first_name', 'last_name', 'email', 'role_id') +
             ['password' => Hash::make(self::DEFAULT_PASSWORD)]
@@ -61,6 +68,8 @@ class UserController extends Controller
      */
     public function update(UserUpdateRequest $request, int $id): Response
     {
+        Gate::authorize('edit', 'users');
+
         $user = User::find($id);
         $user->update($request->only('first_name', 'last_name', 'email', 'role_id'));
 
@@ -73,6 +82,8 @@ class UserController extends Controller
      */
     public function destroy(int $id): Response
     {
+        Gate::authorize('edit', 'users');
+
         User::destroy($id);
 
         return response(null, Response::HTTP_NO_CONTENT);
